+++ odoo/js/functions/common.js
@@ -295,6 +295,57 @@ odoo.define("terminal.functions.Common", function (require) {
                 example:
                     "-o \"{'route': '/jsonrpc', 'method': 'server_version', 'params': {'service': 'db'}}\"",
             });
+            this.registerCommand("metadata", {
+                definition: "View record metadata",
+                callback: this._cmdMetadata,
+                detail: "View record metadata",
+                args: [
+                    "s::m:model::1::The record model",
+                    "i::i:id::1::The record id",
+                ],
+                example: "-m res.partner -i 1",
+            });
+            this.registerCommand("parse_simple_json", {
+                definition: "Parse string simple format to json object",
+                callback: this._cmdParseSimpleJSON,
+                detail: "Parse string simple format to json object. Also accepts native json format.",
+                args: ["j::i:input::1::The input string in simple format"],
+                example: "-i \"KeyA=ValueA keyB='The ValueB'\"",
+            });
+        },
+
+        _cmdParseSimpleJSON: function (kwargs) {
+            this.screen.print(kwargs.input);
+            return Promise.resolve(kwargs.input);
+        },
+
+        _cmdMetadata: function (kwargs) {
+            return new Promise(async (resolve, reject) => {
+                let metadata = {};
+                try {
+                    metadata = (
+                        await rpc.query({
+                            method: "get_metadata",
+                            model: kwargs.model,
+                            args: [[kwargs.id]],
+                            kwargs: {context: this._getContext()},
+                        })
+                    )[0];
+
+                    if (typeof metadata === "undefined") {
+                        this.screen.print(
+                            "Can't found any metadata for the given id"
+                        );
+                    } else {
+                        this.screen.print(
+                            this._templates.render("METADATA", metadata)
+                        );
+                    }
+                } catch (err) {
+                    return reject(err);
+                }
+                return resolve(metadata);
+            });
         },
 
         _cmdRpc: function (kwargs) {