+++ tests/test_chrome.py
@@ -2,6 +2,8 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 import os
 from selenium import webdriver
+from selenium.webdriver.chrome.service import Service
+from selenium.webdriver.chrome.options import Options
 from webdriver_manager.chrome import ChromeDriverManager
 from tests.common import SeleniumTestCase
 
@@ -11,9 +13,10 @@ class ChromeTestCase(SeleniumTestCase):
     def setUpClass(cls):
         cls.ODOO_INSTANCE = os.environ.get('TEST_ODOO_INSTANCE', SeleniumTestCase.ODOO_INSTANCE)
         cls.ODOO_INSTANCE_TYPE = os.environ.get('TEST_ODOO_INSTANCE_TYPE', SeleniumTestCase.ODOO_INSTANCE_TYPE)
-        options = webdriver.ChromeOptions()
+        service = Service(ChromeDriverManager(log_level=0).install())
+        options = Options()
         options.add_extension('./OdooTerminal.zip')
-        cls.browser = webdriver.Chrome(ChromeDriverManager(log_level=0).install(), options=options)
+        cls.browser = webdriver.Chrome(service=service, options=options)
 
     # Test No Terminal
     def test_NoTerminal(self):